cloud-functions/summerize_documents/requirements.txt
functions-framework==3.*
langchain
tiktoken
openai
google-cloud-storage